feat: add form for account (#34)
* Draft of add-account form has been added

* Template of the account form has been modified

* Django forms have been used to make money_account form

* Account balance displaying has been corrected due to Monetary class

* Redirect url after add new account has been changed to show details of new account.

* Basic look of edit account page has been created

* django-bootstrap5 library has been installed  instead of django-bootstrap-v5 and all necessary files has been modified due to that change.

* description field has been set as nonrequired

* Add account form style has been changed

* Add first draft of edit account form

* Form for edit account has been finnished

* Change form labels to polish

* Currency code display bug has been handled

* Success message has been added after delete account

+++ monkeys/monkey_budget/views.py
@@ -1,3 +1,5 @@
+from django.contrib.auth.models import User
+from django.http import JsonResponse
 from django.shortcuts import render, get_object_or_404, redirect
 from .models import MoneyAccount, Transaction, SubTransaction
 from .forms import *
@@ -16,17 +18,38 @@ def show_money_account(request, account_id):
     return render(request, 'account/show_account.html', context)
 
 def add_money_account(request):
+    if request.method == 'POST':
+        form = MoneyAccountForm(request.POST)
+        if form.is_valid():
+            account = form.save(commit=False)
+            account.user_id = User.objects.get(pk=2)
+            account.save()
+            return redirect(f'/monkey-budget/konta/{account.id}')
+    else:
+        form = MoneyAccountForm()
     all_accounts = MoneyAccount.objects.filter(user_id=2)
-    context = {'accounts': all_accounts}
-    return render(request, 'account/add_account.html', context)
+    return render(request, 'account/add_account.html', {'form': form, 'accounts': all_accounts})
 
-def edit_money_account(request):
-    return render(request, 'account/edit_account.html')
+def edit_money_account(request, account_id):
+    chosen_account = get_object_or_404(MoneyAccount, pk=account_id)
+    if request.method == 'POST':
+        form = MoneyAccountForm(request.POST, instance=chosen_account)
+        if form.is_valid():
+            form.save()
+            return redirect(f'/monkey-budget/konta/{chosen_account.id}')
+    else:
+        form = MoneyAccountForm(instance=chosen_account)
+
+    all_accounts = MoneyAccount.objects.filter(user_id=2)
+    context = {'account': chosen_account, 'accounts': all_accounts, 'form': form}
+    return render(request, 'account/edit_account.html', context)
 
 def delete_money_account(request, account_id):
-    account = get_object_or_404(MoneyAccount, id=account_id)
-    account.delete()
-    return redirect('dashboard')
+    if request.method == 'POST':
+        account = get_object_or_404(MoneyAccount, id=account_id)
+        account.delete()
+        return JsonResponse({'success': True, 'message': 'Konto zostało usunięte.'})
+    return JsonResponse({'success': False, 'message': 'Nieprawidłowe żądanie.'})
 
 #zaklepuje poniższe linijki pod transakcje
